+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 All notable changes to this project will be documented in this file.
 
 
+## [1.2.0.0] - 2021-08-20
+### Added
+- Added version detection
+- EzYuzu now remembers previously selected Yuzu Folder Location when launched 
+### Changed
+- Folder Browse dialog 
+### Fixed
+- Visual C++ Redistributable installer no longer forces reboot
+- Improved cleanup process
+- UI tweaks to improve UX
+
+
 ## [1.1.0.0] - 2020-07-22
 ### Added
 - ```New Install``` - now also sets "Optimised GPU Defaults"